feat:마이페이지 회원수정, 회원탈퇴, 세션부분 수정 #64

+++ src/main/java/com/tickerbell/jujuclub/auth/dto/SigninDTO.java
@@ -9,11 +9,10 @@ import java.util.Date;
 @NoArgsConstructor
 @AllArgsConstructor
 public class SigninDTO {
-
-  private int userSeq;
-  private String userId;
-  private String userPw;
-  private String userName;
-  private Date createdAt;
-  private Integer userLevel;
+    private Long userSeq;
+    private String userId;
+    private String userPw;
+    private String userName;
+    private Date createdAt;
+    private Integer userLevel;
 }
